# OpenAPI.yaml
openapi: 3.0.3
info:
  title: EIAE -- Emissions Ingestion and Analysis Engine
  version: 1.0.0
  description: |
    REST API for a multi-tenant emissions monitoring platform.
    Manages organizations, sites, assets, emission readings, rollup aggregates,
    policies, and violation detection.
servers:
  - url: http://localhost:{port}/v1
    variables:
      port:
        default: "8080"
paths:
...
  /orgs/{org_id}/sites:
    get:
      operationId: listSites
      tags: [Sites]
      summary: List sites for an organization
      parameters:
        - $ref: "#/components/parameters/org_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Site"
                  total:
                    type: integer
    post:
      operationId: createSite
      tags: [Sites]
      summary: Create a site
      parameters:
        - $ref: "#/components/parameters/org_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                location:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"

  # ── Asset Types ────────────────────────────────────────────────────────
...
  /orgs/{org_id}/sites/{site_id}/readings:
    post:
      operationId: ingestReadings
      tags: [Readings]
      summary: Batch-insert emission readings
      parameters:
        - $ref: "#/components/parameters/org_id"
        - $ref: "#/components/parameters/site_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [readings]
              properties:
                readings:
                  type: array
                  items:
                    type: object
                    required: [asset_id, recorder_device_id, reading, recorded_at]
                    properties:
                      asset_id:
                        type: string
                        format: uuid
                      recorder_device_id:
                        type: string
                        format: uuid
                      reading:
                        type: number
                      unit:
                        type: string
                        default: kg_co2e
                      recorded_at:
                        type: string
                        format: date-time
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  batch_id:
                    type: string
                    format: uuid
                  count:
                    type: integer

  # ── Emissions (Rollups) ────────────────────────────────────────────────
...
  /orgs/{org_id}/policies:
    get:
      operationId: listPolicies
      tags: [Policies]
      summary: List emission policies for an organization
      parameters:
        - $ref: "#/components/parameters/org_id"
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EmissionPolicy"
                  total:
                    type: integer
    post:
      operationId: createPolicy
      tags: [Policies]
      summary: Create an emission policy
      parameters:
        - $ref: "#/components/parameters/org_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entity_type, entity_id, emission_limit, period]
              properties:
                entity_type:
                  type: string
                  enum: [org, site, asset]
                entity_id:
                  type: string
                  format: uuid
                emission_limit:
                  type: number
                unit:
                  type: string
                  default: kg_co2e
                period:
                  type: string
                  enum: [hour, day, month]
                effective_from:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmissionPolicy"
  /policies/{policy_id}/retire:
    post:
      operationId: retirePolicy
      tags: [Policies]
      summary: Retire an emission policy (set effective_to)
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                effective_to:
                  type: string
                  format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

  # ── Violations ─────────────────────────────────────────────────────────
  /orgs/{org_id}/violations:
    get:
      operationId: listViolations
      tags: [Violations]
      summary: List violations for an organization
      parameters:
        - $ref: "#/components/parameters/org_id"
        - name: open_only
          in: query
          schema:
            type: boolean
            default: true
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EmissionViolation"
                  total:
                    type: integer
  /violations/{violation_id}/acknowledge:
    post:
      operationId: acknowledgeViolation
      tags: [Violations]
      summary: Acknowledge a violation
      parameters:
        - name: violation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
components:
  parameters:
    org_id:
      name: org_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    site_id:
      name: site_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 200
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  schemas:
...
    Site:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        location:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    EmissionPolicy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        entity_type:
          type: string
          enum: [org, site, asset]
        entity_id:
          type: string
          format: uuid
        emission_limit:
          type: number
        unit:
          type: string
        period:
          type: string
          enum: [hour, day, month, year]
        effective_from:
          type: string
          format: date-time
        effective_to:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    EmissionViolation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        entity_type:
          type: string
          enum: [org, site, asset]
        entity_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
        period:
          type: string
        limit_value:
          type: number
        unit:
          type: string
        period_start:
          type: string
          format: date-time
        measured_value:
          type: number
        overage:
          type: number
        detected_at:
          type: string
          format: date-time
        last_evaluated_at:
          type: string
          format: date-time
        acknowledged_at:
          type: string
          format: date-time
          nullable: true